refactor:disconnected-page를 공용 페이지로 조정한다.
