--- apps/web/src/app/products/[productID]/reviews/disconnected-page.tsx
@@ -1,8 +1,0 @@
-export default function DisconnectedPage() {
-  return (
-    <main>
-      <h1>Disconnected</h1>
-      <p>Not connected to a shop</p>
-    </main>
-  );
-}
